+++ NEWS.md
@@ -1,3 +1,7 @@
+# SSVS 2.1.0 (2025-02-03)
+
+- Added functions `SSVS_MI()`, `plot_MI()`, and `summary_MI()` for SSVS with multiply imputed data
+- Added an example demonstrating SSVS for multiply imputed data
 
 # SSVS 2.0.0 (2022-05-27)